backend1/requirements.txt
flask
flask-cors
requests
numpy
transformers
torch
gunicorn
yfinance
python-dotenv
openai